Issue #181
CHANGE:
  - Moved all `SQLStore` stuff into `database/impl/sqlstore.go`. All other
    changes are accompanying this.
  - server/server.go & reaper/reaper.go Now uses
    `database/impl/sqlstore.go:InitSQLStoreByDBChoice()` so that they don't
    need to be aware of individual implementations.
