+++ database/database.go
@@ -1,11 +1,9 @@
 package db
 
 import (
-	"database/sql"
 	"github.com/GrappigPanda/notorious/config"
 	"github.com/GrappigPanda/notorious/database/mysql"
 	"github.com/GrappigPanda/notorious/database/postgres"
-	"github.com/GrappigPanda/notorious/database/schemas"
 	"github.com/jinzhu/gorm"
 	// We use a blank import here because I'm afraid of breaking anything
 	_ "github.com/jinzhu/gorm/dialects/mysql"
@@ -29,39 +27,14 @@ func InitDB(config *config.ConfigStruct) {
 	}
 }
 
-// PeerDeltaEvent allows us to set an event to handle how we're going to update
-// teh database.
-type PeerDeltaEvent int
+func OpenDBChoiceConnection() (*gorm.DB, error) {
+	cfg := config.LoadConfig()
 
-const (
-	// PEERUPDATE represents a change to a peer, so we'll update a tracker
-	// user's ratio.
-	PEERUPDATE PeerDeltaEvent = iota
-	// TRACKERUPDATE handles updating total tracker stats
-	TRACKERUPDATE
-	// TORRENTUPDATE represents the changes to a specific torrent where we
-	// update total upload/download for the torrent itself.
-	TORRENTUPDATE
-)
-
-// PeerTrackerDelta handles holding data to be updated by the `UpdateConsumer`.
-type PeerTrackerDelta struct {
-	Uploaded   uint64
-	Downloaded uint64
-	IP         string
-	Event      PeerDeltaEvent
-}
-
-// SQLStore is the base implementation for a database which will be used to
-// store stats and retrieve whitelisted torrents.
-type SQLStore interface {
-	OpenConnection() (*gorm.DB, error)
-	GetTorrent(string) (*schemas.Torrent, error)
-	GetWhitelistedTorrent(string) (*schemas.WhiteTorrent, error)
-	UpdateStats(uint64, uint64)
-	UpdateTorrentStats(int64, int64)
-	ScrapeTorrent(string) *schemas.Torrent
-	GetWhitelistedTorrents() (*sql.Rows, error)
-	UpdatePeerStats(uint64, uint64, string)
-	HandlePeerUpdates() chan PeerTrackerDelta
+	if cfg.DBChoice == "mysql" {
+		return mysql.OpenConnection()
+	} else if cfg.DBChoice == "postgres" {
+		return postgres.OpenConnection()
+	} else {
+		panic("Invalid database choice found for `OpenDBChoiceConnection`.")
+	}
 }
